+++ packages/bitcoin-module/src/handlers/get-network-fee/get-network-fee.ts
@@ -1,0 +1,34 @@
+import type { NetworkFees } from '@avalabs/vm-module-types';
+
+import { getProvider } from '../../utils/get-provider';
+
+/**
+ * Returns {@link NetworkFees} based on `estimatesmartfee` RPC call
+ */
+export async function getNetworkFee({
+  isTestnet,
+  proxyApiUrl,
+}: {
+  isTestnet: boolean;
+  proxyApiUrl: string;
+}): Promise<NetworkFees> {
+  const provider = getProvider({
+    isTestnet,
+    proxyApiUrl,
+  });
+
+  const { high, low, medium } = await provider.getFeeRates();
+
+  return {
+    low: {
+      maxFeePerGas: BigInt(low),
+    },
+    medium: {
+      maxFeePerGas: BigInt(medium),
+    },
+    high: {
+      maxFeePerGas: BigInt(high),
+    },
+    isFixedFee: false,
+  };
+}