+++ packages/bitcoin-module/src/module.test.ts
@@ -1,0 +1,29 @@
+import { Environment, type Network } from '@avalabs/vm-module-types';
+
+import { BitcoinModule } from './module';
+import { getNetworkFee } from './handlers/get-network-fee';
+import { devEnv, prodEnv } from './env';
+
+jest.mock('./handlers/get-network-fee');
+
+describe('bitcoin-module', () => {
+  describe('getNetworkFee()', () => {
+    it('uses the get-network-fee handler', async () => {
+      const devModule = new BitcoinModule({ environment: Environment.DEV });
+      await devModule.getNetworkFee({ isTestnet: true } as Network);
+
+      expect(getNetworkFee).toHaveBeenCalledWith({
+        isTestnet: true,
+        proxyApiUrl: devEnv.proxyApiUrl,
+      });
+
+      const prodModule = new BitcoinModule({ environment: Environment.PRODUCTION });
+      await prodModule.getNetworkFee({ isTestnet: false } as Network);
+
+      expect(getNetworkFee).toHaveBeenCalledWith({
+        isTestnet: false,
+        proxyApiUrl: prodEnv.proxyApiUrl,
+      });
+    });
+  });
+});