+++ packages/evm-module/src/handlers/eth-send-transaction/eth-send-transaction.test.ts
@@ -84,7 +84,7 @@ const displayData = {
 };
 
 const signingData = {
-  type: 'evm_transaction',
+  type: 'eth_sendTransaction',
   account: '0xfrom',
   chainId: 1,
   data: {