+++ packages/types/src/network-fee.ts
@@ -2,6 +2,6 @@ export type NetworkFees = {
   low: { maxFeePerGas: bigint; maxPriorityFeePerGas?: bigint };
   medium: { maxFeePerGas: bigint; maxPriorityFeePerGas?: bigint };
   high: { maxFeePerGas: bigint; maxPriorityFeePerGas?: bigint };
-  baseFee: bigint;
+  baseFee?: bigint;
   isFixedFee: boolean;
 };